backend/requirements.txt
fastapi
uvicorn
pypdf
python-docx
openai
python-multipart
python-dotenv
